Changes to comments on 12 Feb 1. Rename home.py to home_locations.py 2. Rename location.py to query_accommodation_coords.py 3. Move build_households_home_list() from city.py to obtain_accommodation_coords.ipynb, save full homes specs list to full_home_list.csv 4. Add get_home_samples() func to home_location.py to retrieve home samples list from full_home_list.csv 5. Change the call of build_households_home_list() to the call of get_home_samples() in city.py 6. Remove reset_people() from city.py, instead change the reset_people() in population.py 7. Construct the Home once per household in city.py 8. Change constant POPULATION_CITY and CITY_AREA into two key-value pairs as an item of the dictionary city_paras{}, which can save multiple regions parametres 9. Add print_city_paras() in city_cfg.py
